Separate shared setup from e2e test
